sql: add sequence option info for identity columns under information_schema
Previously, for a column created with the `GENERATED ... AS IDENTITY
(seq_options)` syntax, the info for the sequence option is not saved in the
information schema.

This commit is to fix it. We parse the sequence options saved as a string in
the descriptor, so that it's much easier to extract specific option such as
sequence's start value or increment size.

To make sure that we get the same sequence option to generate
the sequence, we reuse `assignSequenceOptions()` by breaking it into several
helper functions.

fixes https://github.com/cockroachdb/cockroach/issues/82064

Release note (sql change): add sequence option info for identity columns under
information_schema
