+++ pkg/sql/catalog/schemaexpr/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "hash_sharded_compute_expr.go",
         "partial_index.go",
         "select_name_resolution.go",
+        "sequence_options.go",
         "unique_contraint.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/catalog/schemaexpr",
