+++ pkg/sql/catalog/schemaexpr/column.go
@@ -107,7 +107,7 @@ func FormatColumnForDisplay(
 				f.WriteString(" GENERATED BY DEFAULT AS IDENTITY")
 			}
 			if col.HasGeneratedAsIdentitySequenceOption() {
-				seqOpt := col.GetGeneratedAsIdentitySequenceOption()
+				seqOpt := col.GetGeneratedAsIdentitySequenceOptionStr()
 				s := formatGeneratedAsIdentitySequenceOption(seqOpt)
 				f.WriteString(s)
 			}
